Add option to append the question to the end of the context
If `--append_question_to_context_too` flag is set during training, batch evaluation or server mode, the question will get copied to
the end of context with an additional space in between. The context will then be tokenized like a non-program field.
This option's main purpose is to be used with pretrained language models and use their tokens/word embeddings as much as possible.