numericalizer: only use fast tokenizers that are known to work
Unfortunately, not all fast tokenizers work: some ignore
do_basic_tokenize, so they mess up the spacing/symbols in the input,
and some mess up the spacing around special tokens when reversing.

Add a list of models that are known to work with fast tokenizers.
Everything else is slow until tested and known to work.